[not verified] # This is a combination of 2 commits.
We don't need to enqueue Swiper via wp_enqueue_script since we're importing the module.
Abstract xhr
